move Feature type to types dir
